requirements.txt
# Flask & Web
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
pg8000==1.30.3
alembic==1.13.1

# Celery & Message Queue
celery==5.3.4
kombu==5.3.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Data Processing & ML
numpy>=2.4.2
scipy>=1.17.0
scikit-learn>=1.8.0
python-dateutil==2.8.2
metaflow==2.19.18
xgboost==3.1.3
joblib==1.5.3

# Testing
pytest==7.4.3psycopg2-binary==2.9.9
pillow==10.1.0
